refactor(mobile): view init (#4263)
